output/project_run_20250526170838/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
Flask-CORS
requests
pandas
openpyxl
pytest
pytest-mock
coverage
Faker
Alembic